Add RepeatTypeEnum, update TaskModel interface, and refactor taskTrackerTable
Added RepeatTypeEnum to the list of exported enums in the backend-configuration-pn/enums/index.ts file. Created a new repeat-type.enum.ts file that exports a RepeatTypeEnum enum with values for Day, Week, and Month. Updated the TaskModel interface in the backend-configuration-pn/models/task-tracker/task.model.ts file to use RepeatTypeEnum instead of number for the repeatType property. Refactored the task-tracker-table component HTML in order to display weeks and days more efficiently when viewing tasks on a calendar. Replace calculate weeks, days and etc to helper (after get data from db)