Add logic to task wizard page
The commit adds logic to the task wizard page by introducing a new helper function called "generateWeeksList". This function generates a list of weeks up to a specified limit and is used in the component "AreaRulePlanModalComponent" to populate the "repeatEveryArr" array. Additionally, two new enum files are added: "task-wizard-statuses.enum.ts" and "index.ts". The translation files for different languages are also updated with new translations related to tasks.