Try fix tests
Fixes an issue with the task-wizard.edit.spec.cy.ts file where the folder creation process was not working correctly. Added a wait time before entering the folder name and description to ensure proper loading. Also added a wait time before clicking the save button to allow for successful saving of the folder.