Try fix tests
Fixes an issue where the wrong button was being clicked in the task wizard update modal. Also updates the selection of values in the updateTableTags and updateTags elements. Additionally, adds functionality to enable/disable tasks for editing names. Finally, verifies that the table displays the correct information after updating a task.